feat: generic runtime error
