+++ crates/pop-drink/src/lib.rs
@@ -5,7 +5,7 @@ use ink_sandbox::{AccountIdFor, BalanceFor};
 pub use session::{error::SessionError, ContractBundle, Session, NO_SALT};
 pub use sp_io::TestExternalities;
 
-pub mod errors;
+pub mod error;
 #[cfg(test)]
 pub mod mock;
 pub mod utils;
@@ -17,16 +17,28 @@ pub mod devnet {
 
 	/// A set of primitive runtime errors and versioned runtime errors used for testing.
 	pub mod error {
-		pub use pop_runtime_devnet::RuntimeError::{self as PopRuntimeError, *};
+		pub use pop_runtime_devnet::RuntimeError::*;
 
-		pub use crate::errors::*;
+		pub use crate::error::*;
+
 		/// A collection of error types from the `v0` module used for smart contract testing in the
 		/// `devnet` environment.
 		pub mod v0 {
 			use super::*;
-			pub use crate::errors::v0::*;
+			pub use crate::error::v0::*;
+
+			/// Configuration for the runtime error used to test smart contract.
+			#[derive(Debug)]
+			pub struct Config;
+			impl RuntimeErrorConfig for Config {
+				type ModuleError = pop_runtime_devnet::RuntimeError;
+				type PopApiError = pop_primitives::v0::Error;
+
+				const MODULE_INDEX: u8 = 3;
+			}
+
 			/// Runtime error type used for testing smart contract methods using Pop API `v0`.
-			pub type RuntimeError = crate::errors::v0::RuntimeError<PopRuntimeError>;
+			pub type RuntimeError = crate::error::RuntimeError<Config>;
 		}
 	}
